# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend App Ecommerce API",
        "description": "Documentación de la API del Ecommerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3001",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/direcciones-envio": {
            "get": {
                "tags": [
                    "Direcciones de Envío"
                ],
                "summary": "Listar direcciones de envío",
                "operationId": "5e1663a00f7e52518787417c7c95ca2c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar dirección por alias, dirección, urbanización, sector o referencia",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de direcciones de envío obtenida correctamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Direcciones de Envío"
                ],
                "summary": "Registrar dirección de envío",
                "description": "Crea una nueva dirección de envío asociada a un usuario.",
                "operationId": "59905d3f3b49fc70c110aef5ccb42765",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "direccion",
                                    "es_direccion_principal",
                                    "estado"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "description": "ID del usuario propietario de la dirección",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "alias_direccion": {
                                        "description": "Alias o nombre identificador de la dirección",
                                        "type": "string",
                                        "example": "Casa",
                                        "nullable": true
                                    },
                                    "urbanizacion": {
                                        "description": "Urbanización",
                                        "type": "string",
                                        "example": "Urb. Las Palmas",
                                        "nullable": true
                                    },
                                    "sector": {
                                        "description": "Sector",
                                        "type": "string",
                                        "example": "Sector 2",
                                        "nullable": true
                                    },
                                    "direccion": {
                                        "description": "Dirección principal de envío",
                                        "type": "string",
                                        "example": "Av. Los Olivos 123"
                                    },
                                    "manzana": {
                                        "description": "Manzana",
                                        "type": "string",
                                        "example": "A",
                                        "nullable": true
                                    },
                                    "lote": {
                                        "description": "Lote",
                                        "type": "string",
                                        "example": "15",
                                        "nullable": true
                                    },
                                    "referencia": {
                                        "description": "Referencia para facilitar la ubicación",
                                        "type": "string",
                                        "example": "Frente al parque principal",
                                        "nullable": true
                                    },
                                    "es_direccion_principal": {
                                        "description": "Indica si es la dirección principal: 1 sí, 0 no",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "estado": {
                                        "description": "Estado de la dirección: 1 activo, 0 inactivo",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dirección de envío creada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/direcciones-envio/{direccionEnvio}": {
            "put": {
                "tags": [
                    "Direcciones de Envío"
                ],
                "summary": "Actualizar dirección de envío",
                "description": "Actualiza una dirección de envío existente asociada a un usuario.",
                "operationId": "16f07c5abbaa5ea6274ef4fbb0280ad5",
                "parameters": [
                    {
                        "name": "direccionEnvio",
                        "in": "path",
                        "description": "ID de la dirección de envío",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "direccion",
                                    "es_direccion_principal",
                                    "estado"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "description": "ID del usuario propietario de la dirección",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "alias_direccion": {
                                        "description": "Alias o nombre identificador de la dirección",
                                        "type": "string",
                                        "example": "Casa",
                                        "nullable": true
                                    },
                                    "urbanizacion": {
                                        "description": "Urbanización",
                                        "type": "string",
                                        "example": "Urb. Las Palmas",
                                        "nullable": true
                                    },
                                    "sector": {
                                        "description": "Sector",
                                        "type": "string",
                                        "example": "Sector 2",
                                        "nullable": true
                                    },
                                    "direccion": {
                                        "description": "Dirección principal de envío",
                                        "type": "string",
                                        "example": "Av. Los Olivos 123"
                                    },
                                    "manzana": {
                                        "description": "Manzana",
                                        "type": "string",
                                        "example": "A",
                                        "nullable": true
                                    },
                                    "lote": {
                                        "description": "Lote",
                                        "type": "string",
                                        "example": "15",
                                        "nullable": true
                                    },
                                    "referencia": {
                                        "description": "Referencia para facilitar la ubicación",
                                        "type": "string",
                                        "example": "Frente al parque principal",
                                        "nullable": true
                                    },
                                    "es_direccion_principal": {
                                        "description": "Indica si es la dirección principal: 1 sí, 0 no",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "estado": {
                                        "description": "Estado de la dirección: 1 activo, 0 inactivo",
                                        "type": "integer",
                                        "example": 1,
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dirección de envío actualizada correctamente"
                    },
                    "404": {
                        "description": "Dirección de envío no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}